requirements.txt
# RAG Data Collector — базовые зависимости (без сборки C-расширений под Windows)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
aiofiles>=23.2.0
beautifulsoup4>=4.12.0
playwright>=1.41.0
PyYAML>=6.0.1
python-multipart>=0.0.6
httpx>=0.26.0
langdetect>=1.0.9
tiktoken>=0.5.0
tenacity>=8.2.0
structlog>=24.1.0
asyncpg>=0.29.0

# PDF: извлечение текста для RAG (files[].extracted_text). Достаточно одного.
pdfminer.six>=20221105
# pdfplumber>=0.10.0   # альтернатива, хорош для таблиц
# PyMuPDF>=1.23.0      # быстрее, но на части систем нужны бинарные колеса

# Опционально: солвер reCAPTCHA v2 через аудио (нужен FFmpeg в PATH). При CAPTCHA на КиберЛенинке — pip install playwright-recaptcha && winget install ffmpeg
# playwright-recaptcha>=0.5.0

# Опционально (часто ломают сборку на Windows из-за blis/native):
# spacy>=3.7.0   — альтернативный токенизатор; по умолчанию используется tiktoken
# fasttext-wheel — альтернатива langdetect; по умолчанию используется langdetect
